Make the drum kit work
I had to watch the video to figure out how to stop and restart the
drum sound.

I also had no clue how to get the playing class off the
element after the audio finished. I was on a totally different path and
was trying to do something with a callback from the play() method...
which there might not even be one? 😬