Return collection of anchors for exercise 6
I started toying with the idea of cleaning my data so I wasn't working
with anchors, but that I was just grabbing street names. But then this
felt too cutesy so I cut it out.

Now I'm seeing a big list of anchor elements and I don't like it.
Fixing...