Fix exercise 6 by using anchor text, not anchors
Instead of handling the nodeList, I want to work an array that contains
just the important information. I do this with `Array.prototype.map()`
and pull out just the anchor text. Then I can use filter on that array
instead of monkeying around with the nodeList object that doesn't even
have a map function.