Steal Wes's exercise 7 homework
I think it's a little silly the way ya boy's deconstructing an array
when he's only going to use the first member of it, so I'm being kinda
gross and explicitly grabbing the node with the index of 0 using square
brackets.

I'm wondering where this exercise should take me in terms of how I
should handle the case where two persons have the same last name. I'm
about to optimize to sort by that also, although I don't know if that's
breaking the rules. 🤷🏻‍♂️