Complete exercise 6
Ok, this one was pretty by-the-numbers also. I was exposed to some
things I'm not really familiar with like Fetch API and the RegExp
constructor. Specifically it seems like I always bungle which way it's
supposed to be called -- String.match(regexp) or RegExp.match(string);

I'm not sure exactly how the regular expression replace is working by
replacing the matched expression with identical casing, but I guess
that's not important.