Complete exercise 7
On this one I did sort of a hybrid of painting by the numbers and not
painting by the numbers. I attacked the first problems on my own with
some different solutions that I kinda wish I'd committed (but didn't)
and then played back his solution. I updated to make use of his `new
Date().getFullYear()` thing -- just like random numbers, I don't do a
whole lot of work with the Date API... so I was just like "idk, 2019"

I also wasn't familiar with the `Array.prototype.splice()` method in the
sense that I knew how to use it. Instead I used the delete keyword to
remove that member from the array, and then discovered that's a real
weird way to do it.

All in all a useful exercise! Less fun than the first, but still pretty
cool.