Complete exercise 8
This was more following along with a video, but I felt like it was a
good mix of me just blindly following the video (cause I don't do a lot
of work with `<canvas>` and me coming up with my own solutions for
things.

I kind of like my approach with the "direction" variable where I do
`lineWidth += (direction ? 1 : -1)` instead of breaking out on lines to
grow or shrink... but on the other hand, maybe this is too clever and
would be better expressed with an if-else statement.

LOTS of new stuff here with `<canvas>` but not a ton of new JS-specific
stuff. 🎨