Refactor to use variables more
I like the idea of using `isHalfShown` and `isAboveViewport` etc to
store logic and describe what some of the comparisons are instead of
putting them directly into our `if` statements. At risk of turning
into "My code is self-documenting" guy, this seems like a really useful
way to make it obvious to the next person to work on some code what
exactly a comparison is checking for.