Work through examples of reference vs copy
The short version is that when storing native primitives like bools,
strings, and numbers in variables (using keywords `let`, `var`, or
`const`) the data is passed as a copy. But when storing other data
types like arrays or objects, the data is stored as a reference --
when you update one, you update the other.

In order to get around this if you'd like to update one or the other,
you'll need to explicitly copy the data using things like
`Array.from()` or `Object.assign({}, original)` so you don't suffer
from unexpected side effects.

This kind of exercise is intended to build understanding of fundamental
concepts, but I don't know that it's the best way to do it. I wonder
what a better approach might be -- maybe make some failing tests pass
so you can see visible output? Work the concept into a practical example
that's easy to reason about while still being realistic.

I'll try to remember primitives get copied and others get passed by
reference.