Implement very cool looking page
This I think was my least favorite exercise because a lot of things
were thrown at me that weren't explained in a way where it felt like
I had a chance to come up with the answers.

The lessons to get from this are:
  * Object destructuring's syntax makes no sense
  * Working with the mousemove event can be tricky
  * CSS stuff where the JS does high framerate stuff is neat
  * Formulas are hard for me to follow

I like that it looks really cool, but I didn't feel like I was given
what I needed to arrive at the answer on my own.