Do event listener thing closer to Wes
Maybe my own approach was inefficient or unnecessarily burdensome on the
browser, adding a ton of event listeners. Normally I look at these kinds
of optimizations as ignant, but in a use case like this where you've got
a user potentially hitting a button dozens/hundreds/thousands of times,
I can imagine it bogging down the browser. And in music, timing matters.

I'll go with Wes's apparently safer approach and be thankful he's
offered it. ❤️