Rotate the clock face
Because the styles for the clock's hands lay them from right to left,
Wes recommends having them each start rotated 90deg and managing that
number in the JS. I don't like this approach. Don't rotate the hands and manage a magic number, rotate the whole dang clock. So that's what I done did.