Make the clock show hours and minutes
This was pretty straight forward, just followed what I did for seconds
and applied it for hours and minutes.

One thing I don't like is that little stuttering hand thing, but I
remember last time I did this exercise I took WAAAAY too long on that,
so I'm going to skip it for now. No biggie. I'm imagining the right way
to handle that anyway would be by changing some arcane property having
to do with the transition timing, but Wes advocates for doing something
with `if` statements. Naw, I'm set. 😇