List inventors born in 1500s
Okay, so this works, but it's also unnecessarily difficult to read.
We've got a one-liner that uses destructuring in the function signature.
It works, but it's also stupid -- and goes back to my bad habit of not
writing _testable_ code!

I'm going to see if I can do a little better...