List inventors from earliest to latest birth year
I don't really love my approach here, but the `compareFunction` syntax
was a little hard for me to wrap my head around. I started going down
the path of returning -1, 0, or 1 depending on who was older and using a
switch statement but I determined that would be stupid and confusing.