List people by last name
I took a wrong turn at the start of this one by thinking I had to do a
bunch of work on the inventors object, but then I realized I was
overthinking it when I started watching Wes do it and he was working
with a totally different set of data. Total lay-up! 🏀