Implement flex-box and class toggling
This felt to me more like a CSS exercise than a JS exercise, but I liked
getting the opportunity to get more hands-on with the "transitionend"
event. Kinda neat that it fires a different event for every property
that's transitioned instead of considering all the element's transitions
together as one monolithic event.

I'm a little disappointed that I didn't really even know where to start
on this exercise without seeing Wes take the lead on it, but since I'm
very inexperienced with flexbox now, I think that's okay.