Complete exercise 7
This was a much shorter "array cardio" session than the previous one,
but the methods used seem a lot less handy than the ones I was exposed
to in array cardio 1.

I hope we get a better method for removing nodes from an array in the
future -- lines 58-61 are p gross imho. Why so much ceremony over what I
should be able to simply handle with a non-destructive (i.e., doesn't
modify) Array method?