Finish exercise 8
Lots of new stuff for me -- I'm not familiar with `canvas` but I enjoyed
messing around with the number stuff. I like my approach to
growing/shrinking the line better than what chaboy Wes came up with
because I think it's a lot clearer to use a number to show direction (1
or -1) vs a boolean, then I get to do the ol' `direction *= 1` move to
flip it.

I'm not sure that I like the ternary operator I'm doing on line 51 but
w/e, the comment explains it well enough.