Solve exercise 10 with my own solution
Not bad, although I didn't implement holding shift to uncheck multiple
checkboxes. I think this is a good example of a time where tests can
provide a contract, and you can work with non-technical folks to
determine the assertions ("write the spec") before doing the work.

I think a hard thing is making that spec-writing not take up a gajillion
hours. What's the right way of doing it?