Finish exercise 19
WHOOOO totally caught up with where I left off before.

Okay, this exercise was interesting but it was really more of a deep
dive into using canvas and webcam APIs to do some trickery. Cool, but
totally not JS fundamentals. I liked the part where Wes cycled over the
images with the clever for loop that increments 4.