Complete exercise 22
This wasn't so hard once I had an idea what techniques to use. I wasn't
familiar with `getBoundingClientRect()`, but I immediately knew what was
happening with the scroll thing and fixed that handily. Maybe it
would've been more elegant if I'd made my own object to contain all the
appropriate values instead of working directly on the element's style
attributes, but whatever.