Finish exercise 23
I took a different approach on this one to start with than Wes, where
when the visitor submits the form it gets the value from the fields and
uses those options for speaking... but I didn't know how to make that
work elegantly with the "options" nodeList he was using.

Pretty cool, I think I learned that while I could have hammered this one
out on my own, but holding my horses and watching him make his approach,
I could learn more. I wonder if it maybe would be better for me to do
the exercise on my own first and then see his approach. It'd take longer
but maybe I'd learn even more?

Also this was another Web API tutorial, but it gave me some hands-on
practice with filter and reduce, so that was cool. Also `find()` was
neat to see in action.