Follow along with Wes's example
I did almost everything the exact same way BUUUUT he does something a
little better with putting the class responsible for fixing the nav onto
the body instead of the element I'm trying to fix.

Also I added a little pizzazz with the animation for the logo... but I
wish I knew how to make it only do its thing when scrolling in, not
scrolling out.